Restructure handler APIs
