+++ src/handler/handler.test.ts
@@ -1,6 +1,6 @@
 import { it, describe } from "node:test";
 import * as assert from "node:assert/strict";
-import * as nexus from "./index";
+import * as nexus from "../index";
 
 const myService = nexus.service("service name", {
   syncOp: nexus.operation<string, string>(),
